backend/pyproject.toml
[project]
name = "cover-letter-maker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pymupdf>=1.25.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

